+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Changelog
 
 Unreleased
 ==========
+* feat: Added site field in Snippet model
+* feat: Added site filter in Snippet Admin
 * add support for python 3.10
 * add support for django 4.2
 * drop support for django < 3.2