+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog
 
 Unreleased
 ==========
-* feat: Added site field in Snippet model
+* feat: Added sites support for Snippets
 * feat: Added site filter in Snippet Admin
 * add support for python 3.10
 * add support for django 4.2