+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 
 Unreleased
 ==========
+* feat: Added sites support for Snippets
 * add support for python 3.10
 * add support for django 4.2
 * drop support for django < 3.2